Personaje recarga manualmente y las cartas lanzan balas

+++ project/src/our_scripts/components/Mana.h
@@ -10,8 +10,8 @@ public:
 	__CMPID_DECL__(ecs::cmp::MANA);
 	Mana(int max = 5000, int regen = 750) : _max_mana(max), _mana_regen(regen), _mana_count(0) {}
 	void initComponent() override { _mana_count = 0; }
-	void update(int delta_time) override { // Add override after implementing components 
-		_mana_count = std::min(_mana_count + _mana_regen*delta_time/1000, _max_mana);
+	void update(Uint32 delta_time) override { // Add override after implementing components 
+		_mana_count = std::min(_mana_count + int(_mana_regen*delta_time)/1000, _max_mana);
 		//std::cout << _mana_count << std::endl;
 	}
